Add highlight effect to better differentiate player turns
With this change I'm aiming to improve the user experience by adding a highlight effect so that the user can clearly see whose turn it is at the moment of playing. During the computer's turn, the element will have its opacity reduced to 0.5. When it's the user's turn, the opacity will be set back to 1.